requirements.txt
Pillow==5.1.0
pyocr==0.5.1
PyQt5==5.10.1
sip==4.19.8
six==1.11.0
